// src/types/developer.ts
// src/types/developer.ts
export interface Developer {
  full_name: string;
  title: string;
  bio: string;
  profile_image_url: string;
  location: string;
  email: string;
  phone: string;
  linkedin_url: string;
  github_url: string;
  website_url: string;
  years_experience: string;
  id: string;
  created_date: string;
  updated_date: string;
  created_by_id: string;
  created_by: string;
  is_sample: boolean;
}